requirements.txt
aniso8601==9.0.1
apispec==5.1.1
click==8.0.4
Flask==2.0.3
flask-apispec==0.11.0
Flask-RESTful==0.3.9
importlib-metadata==4.8.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
marshmallow==3.15.0
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
typing_extensions==4.1.1
webargs==8.0.1
Werkzeug==2.0.3
zipp==3.6.0
